more coverage and refactor
